Major overhaul of main stub (Main.cpp) plus other tweaks.
I've made notes on top of Main.cpp and there are '@plek' prefixed
comments here and there for myself and to explain the idea behind some
additions/removals. Also removed things we we'rent going to use (PNG
dumper, too much of a bitch with Rocket for example, and a lot of timing
related code in the main loop which will also be superseded by Rocket).
Also fiddled a lot with the exception handling in an attempt to provide
a clean as possible exit from failure (in PIMPPPLAYER_DEMO mode, I
retained all lean stuff so if undefined it's much like the old 64KB stub
again). Tomorrow I'll try to address the D3D error stuff I raised an
issue about (though it's late I'm tired and maybe there's a hidden
switch to make all that go away :-)). Bedtime.