Removed Metaballs as an Element/Node type and removed Overlay.
It made no sense to be one since it's it handles all transformations and
management by itself now. This may change in the future. It also allowed
me to move the draw-or-not hack to Demo.cpp where it belongs.

Overlays are superseded by the sprite code, also for backgrounds and
thus no longer needed.