Possible support for backgrounds (read desc!).
It's being rendered right before the scene so it will in many cases
require that the right blend mode be set (alphablend) and the shader
that renders on top of it should output pixels accordingly (i.e. black
outside traced geometry has alpha 0). We'll have to tweak this per part.