Fixed up Sprites system.
It's less retarded and more flexible now, uses a single vertex buffer
and you can have an arbitrary amount of both back- and foreground
sprites. Also patches the TPB-06 parts to work properly. Then added a
shelldrop, and a placeholder Platform.h in core (see Issues for my plans
with that).